move around model classes
