+++ localstack/services/awslambda/invocation/lambda_models.py
@@ -1,0 +1,1 @@
+# previous API below
